Maj caisse + résas

+++ htdocs/bimpcaisse/controllers/indexController.php
@@ -129,11 +129,11 @@ class indexController extends BimpController
                     'title'   => 'Clients',
                     'content' => $this->renderClientsTabHtml()
                 ),
-                array(
-                    'id'      => 'stocks',
-                    'title'   => 'Stocks produits',
-                    'content' => 'Ici, pourquoi pas: consulter les stocks produits'
-                )
+//                array(
+//                    'id'      => 'stocks',
+//                    'title'   => 'Stocks produits',
+//                    'content' => 'Ici, pourquoi pas: consulter les stocks produits'
+//                )
             );
 
 
@@ -825,20 +825,25 @@ class indexController extends BimpController
             if ($this->isCaisseValide($caisse, $errors)) {
                 $caisse = BimpObject::getInstance($this->module, 'BC_Caisse', $id_caisse);
                 $vente = BimpObject::getInstance($this->module, 'BC_Vente', $id_vente);
-                $vente->set('status', 1);
-                $vente->set('id_caisse', $id_caisse);
-                $vente->set('id_caisse_session', (int) $caisse->getData('id_current_session'));
-                $vente->set('id_entrepot', (int) $caisse->getData('id_entrepot'));
 
-                $errors = $vente->update();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $vente->set('status', 1);
+                    $vente->set('id_caisse', $id_caisse);
+                    $vente->set('id_caisse_session', (int) $caisse->getData('id_current_session'));
+                    $vente->set('id_entrepot', (int) $caisse->getData('id_entrepot'));
 
-                $data = array();
-                $html = '';
+                    $errors = $vente->update();
 
-                if (!count($errors)) {
-                    $view = new BC_View($vente, 'creation');
-                    $html = $view->renderHtml();
-                    $data = $vente->getAjaxData();
+                    $data = array();
+                    $html = '';
+
+                    if (!count($errors)) {
+                        $view = new BC_View($vente, 'creation');
+                        $html = $view->renderHtml();
+                        $data = $vente->getAjaxData();
+                    }
                 }
             }
         }
@@ -879,22 +884,38 @@ class indexController extends BimpController
 
         $success = '';
         if (!count($errors)) {
-            if ((int) $status === 2) {
-                $success = 'Vente validée avec succès';
-                $validate = (int) $vente->validateVente($validate_errors);
-                if (!$validate) {
-                    $errors[] = 'Cette vente ne peut pas être validée';
-                } else {
-                    $ticket_html = $vente->renderTicketHtml($ticket_errors);
-                }
+            if ($vente->getData('status') === 2) {
+                $errors[] = 'Cette vente ne peut pas être modifiée car elle a été validée';
             } else {
-                if ((int) $status === 1) {
-                    $success = 'Vente enregistrée avec succès';
+                if ((int) $status === 2) {
+//                    if ($vente->getData('status') === 0) {
+//                        $errors[] = 'Cette vente ne peut pas être validée car elle a été annulée';
+//                    } else {
+                    $success = 'Vente validée avec succès';
+                    $validate = (int) $vente->validateVente($validate_errors);
+                    if (!$validate) {
+                        $errors[] = 'Cette vente ne peut pas être validée';
+                    } else {
+                        $ticket_html = $vente->renderTicketHtml($ticket_errors);
+                    }
+                    $validate_errors[] = 'TEST';
+                    if (count($validate_errors)) {
+                        $msg = 'Erreur validation vente ' . $vente->id . "\n";
+                        foreach ($validate_errors as $e) {
+                            $msg .= ' - ' . $e . "\n";
+                        }
+                        dol_syslog($msg, LOG_ERR);
+                    }
+//                    }
                 } else {
-                    $success = 'Vente abandonnée';
+                    if ((int) $status === 1) {
+                        $success = 'Vente enregistrée avec succès';
+                    } else {
+                        $success = 'Vente abandonnée';
+                    }
+                    $vente->set('status', (int) $status);
+                    $errors = array_merge($errors, $vente->update());
                 }
-                $vente->set('status', (int) $status);
-                $errors = array_merge($errors, $vente->update());
             }
         }
 
@@ -931,11 +952,15 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $vente->set('id_client', (int) $id_client);
-                $vente->set('id_client_contact', 0);
-                $errors = $vente->update();
-                if (!count($errors)) {
-                    $html = $vente->renderClientView();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $vente->set('id_client', (int) $id_client);
+                    $vente->set('id_client_contact', 0);
+                    $errors = $vente->update();
+                    if (!count($errors)) {
+                        $html = $vente->renderClientView();
+                    }
                 }
             }
         }
@@ -965,10 +990,14 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $vente->set('id_client_contact', (int) $id_contact);
-                $errors = $vente->update();
-                if (!count($errors)) {
-                    $html = $vente->renderContactView();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $vente->set('id_client_contact', (int) $id_contact);
+                    $errors = $vente->update();
+                    if (!count($errors)) {
+                        $html = $vente->renderContactView();
+                    }
                 }
             }
         }
@@ -1049,8 +1078,12 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $html = $vente->selectArticle($id_object, $object_name, $errors);
-                $vente_data = $vente->getAjaxData();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $html = $vente->selectArticle($id_object, $object_name, $errors);
+                    $vente_data = $vente->getAjaxData();
+                }
             }
         }
 
@@ -1091,15 +1124,19 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $article = BimpObject::getInstance($this->module, 'BC_VenteArticle', (int) $id_article);
-                if (!$article->isLoaded()) {
-                    $errors[] = 'Article non trouvé';
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
                 } else {
-                    $article->set('qty', (int) $qty);
-                    $errors = $article->update();
+                    $article = BimpObject::getInstance($this->module, 'BC_VenteArticle', (int) $id_article);
+                    if (!$article->isLoaded()) {
+                        $errors[] = 'Article non trouvé';
+                    } else {
+                        $article->set('qty', (int) $qty);
+                        $errors = $article->update();
 
-                    $total_ttc = (float) ((int) $article->getData('qty') * (float) $article->getData('unit_price_tax_in'));
-                    $stock = (int) $article->getProductStock((int) $vente->getData('id_entrepot'));
+                        $total_ttc = (float) ((int) $article->getData('qty') * (float) $article->getData('unit_price_tax_in'));
+                        $stock = (int) $article->getProductStock((int) $vente->getData('id_entrepot'));
+                    }
                 }
                 $vente_data = $vente->getAjaxData();
             }
@@ -1136,11 +1173,15 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $article = BimpObject::getInstance($this->module, 'BC_VenteArticle', (int) $id_article);
-                if (!$article->isLoaded()) {
-                    $errors[] = 'Article non trouvé';
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
                 } else {
-                    $errors = $article->delete();
+                    $article = BimpObject::getInstance($this->module, 'BC_VenteArticle', (int) $id_article);
+                    if (!$article->isLoaded()) {
+                        $errors[] = 'Article non trouvé';
+                    } else {
+                        $errors = $article->delete();
+                    }
                 }
                 $vente_data = $vente->getAjaxData();
             }
@@ -1176,9 +1217,13 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $remise = BimpObject::getInstance($this->module, 'BC_VenteRemise', $id_remise);
-                $errors = $remise->delete();
-                $vente_data = $vente->getAjaxData();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $remise = BimpObject::getInstance($this->module, 'BC_VenteRemise', $id_remise);
+                    $errors = $remise->delete();
+                    $vente_data = $vente->getAjaxData();
+                }
             }
         }
 
@@ -1217,7 +1262,11 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $html = $vente->addPaiement($code, $montant, $errors);
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $html = $vente->addPaiement($code, $montant, $errors);
+                }
                 $vente_data = $vente->getAjaxData();
             }
         }
@@ -1253,9 +1302,13 @@ class indexController extends BimpController
             if (!$vente->isLoaded()) {
                 $errors[] = 'Cette vente n\'existe plus';
             } else {
-                $paiement = BimpObject::getInstance($this->module, 'BC_VentePaiement', $id_paiement);
-                $errors = $paiement->delete();
-                $html = $vente->renderPaiementsLines();
+                if ($vente->getData('status') === 2) {
+                    $errors[] = 'Cette vente ne peut pas être modifée car elle a été validée';
+                } else {
+                    $paiement = BimpObject::getInstance($this->module, 'BC_VentePaiement', $id_paiement);
+                    $errors = $paiement->delete();
+                    $html = $vente->renderPaiementsLines();
+                }
                 $vente_data = $vente->getAjaxData();
             }
         }
